refactor: decompose into multiple components
