+++ src/components/Carousel/CarouselHeader.scss
@@ -22,7 +22,7 @@
         width: 240px;
         border-radius: 6px;
         border-radius: 50%;
-        background-color: #fff;
+        background-color: white;
         display: flex;
         justify-content: center;
         align-items: center;
@@ -62,8 +62,8 @@
           display: flex;
           justify-content: center;
           align-items: center;
-          color: #fff;
-          font-family: Poppins;
+          color: white;
+          font-family: Poppins, sans-serif;
           font-size: 2rem;
           font-weight: 900;
           text-transform: uppercase;
@@ -97,9 +97,9 @@
           left: -6rem;
         }
         p {
-          color: #fff;
+          color: white;
           text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-          font-family: Poppins;
+          font-family: Poppins, sans-serif;
           font-size: clamp(1.8rem, 4vw, 3rem);
           width: 16ch;
           font-weight: 900;
