+++ src/components/Footer/Footer.scss
@@ -1,0 +1,80 @@
+// Footer
+footer {
+  position: relative;
+  padding: 0;
+  width: 100%;
+  color: white;
+  .footer-top {
+    background-color: #095262;
+    padding-block: 1.5rem;
+    text-align: center;
+    p {
+      font-size: 1rem;
+      text-transform: uppercase;
+    }
+    .social-links-wrapper {
+      .social-link-item {
+        img {
+          height: 45px;
+          width: 45px;
+          margin-inline: 0.9rem;
+          filter: drop-shadow(0px 9px 9px rgb(0 0 0 / 0.6));
+          transition: filter 0.25s ease;
+        }
+        &:hover img {
+          filter: drop-shadow(0px 9px 9px rgb(0 0 0 / 0.1));
+        }
+      }
+    }
+  }
+  .footer-bottom {
+    font-size: 0.8rem;
+    font-weight: 300;
+    min-height: 1.75rem;
+    text-align: center;
+    background-color: #013b47;
+    border-top: 1px solid rgba(255, 255, 255, 0.393);
+    border-bottom: none;
+    p {
+      line-height: 1.5rem;
+      margin-bottom: 0;
+    }
+  }
+  .top-page-link {
+    position: absolute;
+    right: 3rem;
+    top: -1.6rem;
+    height: 3.2rem;
+    width: 3.2rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 50%;
+    background-color: #095262;
+    box-shadow: rgba(0, 0, 0, 0.15) 0px 14px 25px,
+      rgba(0, 0, 0, 0.02) 0px -2px 20px, rgba(0, 0, 0, 0.12) 0px 4px 6px,
+      rgba(0, 0, 0, 0.17) 0px 1px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px;
+    z-index: 2;
+    transition: background-color 0.25s ease-out, box-shadow 0.25s ease;
+    &:hover {
+      background-color: #095262;
+      box-shadow: none;
+    }
+    img {
+      height: 1rem;
+      width: 1.5rem;
+    }
+    &::after {
+      content: '';
+      position: absolute;
+      inset: -3px;
+      border-radius: 50%;
+      border: 4px solid #ffffff00;
+      transition: border-color 0.25s ease-in-out;
+      transition-delay: 0.25s;
+    }
+    &:hover::after {
+      border-color: #f2f5f8;
+    }
+  }
+}
