+++ src/components/Navigation/Navigation.scss
@@ -1,14 +1,10 @@
 nav {
-  box-shadow: rgba(0, 0, 0, 0.15) 0px 2px 8px;
+  box-shadow: rgba(0, 0, 0, 0.35) 0px 6px 6px;
   font-weight: bold;
   .navbar-brand {
     display: flex;
     align-items: center;
     .brand-name {
-      // background: linear-gradient(275deg, #48b6ce 1.7%, #f2d210 121.2%);
-      // background-clip: text;
-      // -webkit-background-clip: text;
-      // -webkit-text-fill-color: transparent;
       color: #00708a;
       font-size: 2rem;
       display: none;
@@ -133,7 +129,7 @@ nav {
   .dropdown-menu.show {
     overflow-y: hidden;
     position: relative;
-    background: rgba(241, 241, 241, 0.616);
+    background-color: rgba(241, 241, 241, 0.616);
     box-shadow: rgba(0, 0, 0, 0.25) -4px 9px 25px -6px;
     border-inline: none;
     border-bottom: 2px solid rgba(255, 255, 255, 0.904);
@@ -172,7 +168,7 @@ nav {
   }
   .navbar-nav .category-item-wrapper {
     background-color: rgb(255, 255, 255);
-    background: linear-gradient(
+    background-image: linear-gradient(
       to right,
       rgb(255, 255, 255) 60%,
       rgba(255, 255, 255, 0.255)
