+++ src/pages/Home/CarouselCategories/CarouselCategories.scss
@@ -66,7 +66,7 @@
         p {
           color: white;
           font-size: 1.8rem;
-          font-family: Poppins;
+          font-family: Poppins, sans-serif;
           font-weight: 500;
           line-height: 111%;
           text-align: center;
