+++ src/pages/Home/PizzaDiscountBanner/PizzaDiscountBanner.scss
@@ -1,0 +1,166 @@
+// Section pizza promotions
+section.section-pizza-discount {
+  margin-top: 8rem;
+  .row {
+    margin-inline: 0;
+    background-color: #00708a;
+    border-radius: 6px;
+    @media screen and (min-width: 1400px) {
+      background-color: transparent;
+    }
+    .section-pizza-discount__left {
+      height: auto;
+      position: relative;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      padding-inline: 0;
+      @media screen and (min-width: 992px) {
+        height: 22.75rem;
+      }
+      .arrow-left-wrapper {
+        position: relative;
+        display: none;
+        @media screen and (min-width: 1400px) {
+          display: block;
+          bottom: 0;
+          left: 0;
+          height: 100%;
+          width: 100%;
+          background-color: transparent;
+          overflow: hidden;
+          border-radius: 6px;
+          z-index: 10;
+          .arrow-left-part-1,
+          .arrow-left-part-2,
+          .arrow-left-part-3 {
+            position: absolute;
+            top: 0px;
+            right: 0;
+            height: 22.75rem;
+            width: 22.75rem;
+            transform-origin: center;
+            background-color: #00708a;
+            border-radius: 6px;
+          }
+          .arrow-left-part-2,
+          .arrow-left-part-3 {
+            box-shadow: rgba(0, 0, 0, 0.85) 0px 25px 50px -12px;
+          }
+          .arrow-left-part-1 {
+            transform: translate(-30%, 0%) rotate(45deg);
+          }
+          .arrow-left-part-2 {
+            transform: translate(-50%, 0%) rotate(45deg);
+          }
+          .arrow-left-part-3 {
+            transform: translate(-110%, 0%) rotate(45deg) scale(1.5);
+          }
+        }
+      }
+    }
+    .content-wrapper {
+      position: relative;
+      left: 0;
+      z-index: 10;
+      text-align: center;
+      padding-top: 2rem;
+      padding-bottom: 0;
+      width: 100%;
+      @media screen and (min-width: 992px) {
+        text-align: left;
+        padding-top: 4rem;
+        padding-left: 8rem;
+        padding-bottom: 2rem;
+      }
+      @media screen and (min-width: 1400px) {
+        position: absolute;
+        left: 3rem;
+        top: calc(50% - 2rem);
+        transform: translateY(-50%);
+        padding-left: 2rem;
+        max-width: 15rem;
+      }
+      p {
+        color: white;
+        line-height: 93%;
+        text-transform: uppercase;
+      }
+      p.title,
+      p.subtitle {
+        font-family: Poppins, sans-serif;
+        font-weight: 900;
+        text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+      }
+      p.title {
+        font-size: clamp(1.8rem, 4vw, 3rem);
+        &.strong {
+          color: #feb95f;
+        }
+      }
+      p.subtitle {
+        font-size: clamp(1.8rem, 2vw, 2.1rem);
+        margin-bottom: 0.75em;
+      }
+    }
+    .section-pizza-discount__right {
+      position: relative;
+      padding-inline: 0;
+      .arrow-right-wrapper {
+        display: none;
+        @media screen and (min-width: 1400px) {
+          display: block;
+          position: relative;
+          bottom: 100%;
+          right: -50%;
+          height: 100%;
+          width: 50%;
+          background-color: #00708a;
+          overflow: hidden;
+          z-index: 10;
+          padding-inline: 0;
+          border-radius: 6px;
+          .arrow-right-part-1,
+          .arrow-right-part-2 {
+            position: absolute;
+            top: 50%;
+            height: 22.75rem;
+            width: 22.75rem;
+            background-color: #00708a;
+            border-radius: 6px;
+          }
+          .arrow-right-part-1 {
+            left: 50%;
+            box-shadow: rgba(0, 0, 0, 0.35) 0px 25px 50px -12px;
+            transform: translate(-82%, -50%) rotate(-45deg);
+          }
+          .arrow-right-part-2 {
+            right: 0;
+            background-color: white;
+            transform: translate(-50%, -50%) rotate(-45deg);
+          }
+        }
+      }
+    }
+    .pizza-wrapper {
+      height: 22.75rem;
+      width: 100%;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      @media screen and (min-width: 1400px) {
+        justify-content: start;
+      }
+      img {
+        position: relative;
+        height: 71%;
+        width: auto;
+        padding-left: 0;
+        z-index: 50;
+        @media screen and (min-width: 1400px) {
+          padding-left: 5%;
+        }
+      }
+    }
+  }
+}
